backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
pydantic[email]==2.4.2
pydantic-settings==2.1.0

# OAuth
authlib==1.3.0
httpx==0.26.0

# Email
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0

# File parsing
pandas==2.1.4
openpyxl==3.1.2

# Rate Limiting
slowapi==0.1.9

# CORS & Middleware

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
